POC for asynchronous process_sign_request2() in ssh-agent
This is probably unsafe, but seems to be functional and relieves a
bottleneck on ssh to several hosts simultaneously by leveraging multiple
cpus on the host running the agent.

The children make no attempt to copy any state back up to the parent.
I was surprised this was unnecessary, even for the basic case I tested.